fix: useSuspenseQuery 도입 페이지의 빌드 타임 prerender fetch 차단
CI 는 더미 API 로 빌드하므로 빌드 중 서버 fetch 가 나가면 prerender 가
실패한다. 이 PR 이 static 라우트에 도입한 useSuspenseQuery 는 서버 prerender
에서 queryFn 을 실제 실행해 API 를 호출하는데, 인증/데이터 페이지들은
클라이언트에서만 fetch 해야 한다.

기존 mypage/credit 의 dynamic(ssr:false) 패턴을 따라 클라이언트 전용으로 전환:
- mypage/application, mypage/career/record: Content 를 도메인 모듈로 분리 후
  dynamic(ssr:false) 로 로드
- 홈 데이터 섹션(CurrentBlog/Interview/Review), 리뷰 ProgramInterviewSection:
  얇은 client 래퍼(dynamic ssr:false)로 감싸 RSC 에서 SSR 제외
  (SEO 는 main 과 동일 — 기존에도 클라이언트 fetch였음)

dynamic 라우트(library/[id], live/[id])는 prerender 대상이 아니라 무변경.
검증: 더미 API 로 next build 통과 (Generating static pages 59/59), ESLint 0.
