+++ apps/web/src/app/(user)/mypage/application/page.tsx
@@ -1,103 +1,16 @@
 'use client';
 
-import { mypageApplicationsQueryOptions } from '@/api/application';
 import { AsyncBoundary } from '@/common/boundary/AsyncBoundary';
 import LoadingContainer from '@/common/loading/LoadingContainer';
-import { CategoryTabs } from '@letscareer/ui';
-import {
-  APPLICATION_CATEGORY_OPTIONS,
-  ApplicationCategory,
-} from '@/domain/mypage/application/constants';
-import ApplySection from '@/domain/mypage/application/section/ApplySection';
-import CompleteSection from '@/domain/mypage/application/section/CompleteSection';
-import EmptySection from '@/domain/mypage/application/section/EmptySection';
-import GuidebookSection from '@/domain/mypage/application/section/GuidebookSection';
-import LibrarySection from '@/domain/mypage/application/section/LibrarySection';
-import ParticipateSection from '@/domain/mypage/application/section/ParticipateSection';
-import VodClassSection from '@/domain/mypage/application/section/VodClassSection';
-import { useSuspenseQuery } from '@tanstack/react-query';
-import { useState } from 'react';
+import dynamic from 'next/dynamic';
 
-const ApplicationContent = () => {
-  const { data: applications } = useSuspenseQuery(
-    mypageApplicationsQueryOptions,
-  );
-  const [category, setCategory] = useState<ApplicationCategory>('PROGRAM');
-
-  const programApplications =
-    applications?.filter(
-      (application) =>
-        application.programType !== 'GUIDEBOOK' &&
-        application.programType !== 'VOD',
-    ) ?? [];
-  const programWaitingList = programApplications.filter(
-    (application) => application.programStatusType === 'PREV',
-  );
-  const programInProgressList = programApplications.filter(
-    (application) => application.programStatusType === 'PROCEEDING',
-  );
-  const programCompletedList = programApplications.filter(
-    (application) => application.programStatusType === 'POST',
-  );
-
-  const guidebookApplicationList =
-    applications?.filter(
-      (application) => application.programType === 'GUIDEBOOK',
-    ) ?? [];
-
-  const vodClassApplicationList =
-    applications?.filter((application) => application.programType === 'VOD') ??
-    [];
-
-  const isProgramEmpty =
-    programWaitingList.length === 0 &&
-    programInProgressList.length === 0 &&
-    programCompletedList.length === 0;
-
-  return (
-    <main className="flex w-full flex-col gap-8 md:gap-10">
-      <div className="-mx-5 -mt-[18px] md:mx-0 md:mt-0">
-        <CategoryTabs
-          options={APPLICATION_CATEGORY_OPTIONS}
-          selected={category}
-          onChange={setCategory}
-        />
-      </div>
-      <div className="flex w-full flex-col gap-16">
-        {category === 'PROGRAM' && (
-          <>
-            {isProgramEmpty ? (
-              <EmptySection
-                text="아직 신청한 프로그램이 없어요"
-                href="/program"
-                buttonText="프로그램 둘러보기"
-              />
-            ) : (
-              <>
-                <ApplySection
-                  applicationList={programWaitingList}
-                  hasInProgress={programInProgressList.length > 0}
-                  hasCompleted={programCompletedList.length > 0}
-                />
-                <ParticipateSection applicationList={programInProgressList} />
-                <CompleteSection applicationList={programCompletedList} />
-              </>
-            )}
-          </>
-        )}
-
-        {category === 'LIBRARY' && <LibrarySection />}
-
-        {category === 'GUIDEBOOK' && (
-          <GuidebookSection applicationList={guidebookApplicationList} />
-        )}
-        {category === 'VOD' && (
-          <VodClassSection applicationList={vodClassApplicationList} />
-        )}
-      </div>
-    </main>
-  );
-};
+// 신청 현황은 클라이언트에서만 fetch 하는 인증 데이터다.
+// ssr:false 로 서버 렌더(빌드 타임 prerender)에서 제외해야
+// 내부 useSuspenseQuery 가 빌드 중 실행돼 API 를 호출(prerender 에러)하지 않는다.
+const ApplicationContent = dynamic(
+  () => import('@/domain/mypage/application/ApplicationContent'),
+  { ssr: false },
+);
 
 const Application = () => {
   return (
