+++ apps/web/src/domain/home/HomeScreen.tsx
@@ -9,13 +9,13 @@ import LogoPlaySection from '@/domain/home/banner/LogoPlaySection';
 import MainBannerSection from '@/domain/home/banner/MainBannerSection';
 import MainCurationSection from '@/domain/home/banner/MainCurationSection';
 import BlogCurationSection from '@/domain/home/blog/BlogCurationSection';
-import CurrentBlogSection from '@/domain/home/blog/CurrentBlogSection';
-import InterviewSection from '@/domain/home/blog/InterviewSection';
+import CurrentBlogSection from '@/domain/home/blog/CurrentBlogSectionClient';
+import InterviewSection from '@/domain/home/blog/InterviewSectionClient';
 import IntroSection from '@/domain/home/Intro/IntroSection';
 import ActiveProgramSection from '@/domain/home/program/ActiveProgramSection';
 import LetsCareerSection from '@/domain/home/program/LetsCareerSection';
 import ReviewCurationSection from '@/domain/home/review/ReviewCurationSection';
-import ReviewSection from '@/domain/home/review/ReviewSection';
+import ReviewSection from '@/domain/home/review/ReviewSectionClient';
 
 export default function HomeScreen() {
   return (
